single-label/requirements.txt
torch>=2.0.0
transformers>=4.30.0
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.2.0
pyyaml>=6.0
tqdm>=4.65.0
accelerate>=0.20.0
bitsandbytes>=0.41.0
datasets>=2.0.0
